Update SkyCommand.ino
Renamed the time variable to avoid name conflicts during compilation. Correspondingly, updated the timeout variable's name to match the new identifier for consistency.

Modified setValue to be a void function due to a compilation error.

Added return values for itsbeentoolong() and removed the shutdown call within that function, as shutdown is handled within the conditional logic in the loop() function

uint8_t can only store values between 0 and 255, removed "value >= 0 &&" from the if conditions.

Commented out Renamed the time variable to avoid name conflicts during compilation. Correspondingly, updated the timeout variable's name to match the new identifier for consistency.

Modified setValue to be a void function due to a compilation error.

Added return values for itsbeentoolong() and removed the shutdown call within that function, as shutdown is handled within the conditional logic in the loop() function

uint8_t can only store values between 0 and 255, removed "value >= 0 &&" from the if conditions.

 Commented out the condition 'duration < 0' because it is always false for unsigned long variables.